+++ library/core/src/main/java/com/google/android/exoplayer2/trackselection/DefaultTrackSelector.java
@@ -1934,6 +1934,7 @@ public class DefaultTrackSelector extends MappingTrackSelector {
           getAdaptiveAudioTracks(
               selectedGroup,
               formatSupports[selectedGroupIndex],
+              params.maxAudioBitrate,
               params.allowAudioMixedMimeTypeAdaptiveness,
               params.allowAudioMixedSampleRateAdaptiveness);
       if (adaptiveTracks.length > 0) {
@@ -1951,6 +1952,7 @@ public class DefaultTrackSelector extends MappingTrackSelector {
   private static int[] getAdaptiveAudioTracks(
       TrackGroup group,
       int[] formatSupport,
+      int maxAudioBitrate,
       boolean allowMixedMimeTypeAdaptiveness,
       boolean allowMixedSampleRateAdaptiveness) {
     int selectedConfigurationTrackCount = 0;
@@ -1967,6 +1969,7 @@ public class DefaultTrackSelector extends MappingTrackSelector {
                 group,
                 formatSupport,
                 configuration,
+                maxAudioBitrate,
                 allowMixedMimeTypeAdaptiveness,
                 allowMixedSampleRateAdaptiveness);
         if (configurationCount > selectedConfigurationTrackCount) {
@@ -1977,13 +1980,16 @@ public class DefaultTrackSelector extends MappingTrackSelector {
     }
 
     if (selectedConfigurationTrackCount > 1) {
+      Assertions.checkNotNull(selectedConfiguration);
       int[] adaptiveIndices = new int[selectedConfigurationTrackCount];
       int index = 0;
       for (int i = 0; i < group.length; i++) {
+        Format format = group.getFormat(i);
         if (isSupportedAdaptiveAudioTrack(
-            group.getFormat(i),
+            format,
             formatSupport[i],
-            Assertions.checkNotNull(selectedConfiguration),
+            selectedConfiguration,
+            maxAudioBitrate,
             allowMixedMimeTypeAdaptiveness,
             allowMixedSampleRateAdaptiveness)) {
           adaptiveIndices[index++] = i;
@@ -1998,6 +2004,7 @@ public class DefaultTrackSelector extends MappingTrackSelector {
       TrackGroup group,
       int[] formatSupport,
       AudioConfigurationTuple configuration,
+      int maxAudioBitrate,
       boolean allowMixedMimeTypeAdaptiveness,
       boolean allowMixedSampleRateAdaptiveness) {
     int count = 0;
@@ -2006,6 +2013,7 @@ public class DefaultTrackSelector extends MappingTrackSelector {
           group.getFormat(i),
           formatSupport[i],
           configuration,
+          maxAudioBitrate,
           allowMixedMimeTypeAdaptiveness,
           allowMixedSampleRateAdaptiveness)) {
         count++;
@@ -2018,9 +2026,11 @@ public class DefaultTrackSelector extends MappingTrackSelector {
       Format format,
       int formatSupport,
       AudioConfigurationTuple configuration,
+      int maxAudioBitrate,
       boolean allowMixedMimeTypeAdaptiveness,
       boolean allowMixedSampleRateAdaptiveness) {
     return isSupported(formatSupport, false)
+        && (format.bitrate == Format.NO_VALUE || format.bitrate <= maxAudioBitrate)
         && (format.channelCount != Format.NO_VALUE
             && format.channelCount == configuration.channelCount)
         && (allowMixedMimeTypeAdaptiveness