+++ testutils/src/main/java/com/google/android/exoplayer2/testutil/HostActivity.java
@@ -166,7 +166,8 @@ public final class HostActivity extends Activity implements SurfaceHolder.Callba
   public void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
     requestWindowFeature(Window.FEATURE_NO_TITLE);
-    setContentView(getResources().getIdentifier("host_activity", "layout", getPackageName()));
+    setContentView(
+        getResources().getIdentifier("exo_testutils_host_activity", "layout", getPackageName()));
     surfaceView = findViewById(
         getResources().getIdentifier("surface_view", "id", getPackageName()));
     surfaceView.getHolder().addCallback(this);